requirements.txt
Django==2.0.1
django-webpack-loader==0.5.0
djangorestframework==3.7.7
